moved getAccessToken as private method in MaventaService
